Collision class, super for x,y,color

+++ src/Game/Manager/Level/Level.java
@@ -3,7 +3,7 @@ package Game.Manager.Level;
 public interface  Level {
      int getNum();
      void handleAiDifficulty(); //todo type?
-     int timeForBlueMode();
-     int timeForFruit();
+     double timeForBlueMode();
+     double timeForFruit();
 
 }
